test isupper

+++ agarkov.sergey/P4/transformMatrixAPeriph.cpp
@@ -19,9 +19,9 @@ int getMinRange(int a, int b, int c, int d)
 
 void agarkov::transformMatrixAPeriph(int* array, size_t rows, size_t cols)
 {
-  for (int i = 0; i < cols; i++)
+  for (size_t i = 0; i < cols; i++)
   {
-    for (int j = 0; j < rows; j++)
+    for (size_t j = 0; j < rows; j++)
     {
       array[i * cols + j] += getMinRange(i + 1, j + 1, rows - i, cols - j);
     }
